chore(core): portability and warning cleanup across cpu, fileio and system
MSVC-only pragmas guarded with _MSC_VER so the Linux build stays quiet,
signed/unsigned comparison casts in the input port lookups, and removal
of the remaining boundary guard blocks.

+++ aae/aae/fileio/mame_fileio.cpp
@@ -161,7 +161,7 @@ int readint(void* f, UINT32* num)
 	int i;
 
 	*num = 0;
-	for (i = 0; i < sizeof(UINT32); i++)
+	for (i = 0; i < (int)sizeof(UINT32); i++)
 	{
 		unsigned char c;
 
@@ -178,7 +178,7 @@ void writeint(void* f, UINT32 num)
 {
 	int i;
 
-	for (i = 0; i < sizeof(UINT32); i++)
+	for (i = 0; i < (int)sizeof(UINT32); i++)
 	{
 		unsigned char c;
 
@@ -193,7 +193,7 @@ int readword(void* f, UINT16* num)
 	int i, res;
 
 	res = 0;
-	for (i = 0; i < sizeof(UINT16); i++)
+	for (i = 0; i < (int)sizeof(UINT16); i++)
 	{
 		unsigned char c;
 
@@ -211,7 +211,7 @@ void writeword(void* f, UINT16 num)
 {
 	int i;
 
-	for (i = 0; i < sizeof(UINT16); i++)
+	for (i = 0; i < (int)sizeof(UINT16); i++)
 	{
 		unsigned char c;
